pyproject.toml
[tool.poetry]
name = "stock-control"
version = "0.1.0"
description = ""
authors = ["maxrangesft <maxrangesft@gmail.com>"]
readme = "README.md"
packages = [{include = "stock_control"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.7"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
